Integrated Local Network Graph + Centrality into index.py
